refactor(ui): split useChatStore.tsx into hooks/chat/ directory
